requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Werkzeug==3.0.1
python-dotenv==1.0.0
bcrypt==4.1.2
